Updated for MSigDB 7.2. Changed to use a dedicated Docker container.
Changed to use a new dedicated Docker container.  The corresponding
Dockerfile is in the gsea-desktop repo.  The build has been changed
since we no longer need to download a GSEA bundle to be included in the
module's ZIP bundle (still need to update this to deal with local test
builds, however).
The MSigDB 7.2 update is very much like the usual, except some new
collections have been added.